etl/requirements.txt
requests
pyyaml
shapely
pyproj
astral>=2,<4
Pillow
scikit-image
